ShadowRansom/requirements.txt
pycryptodome==3.16.0